# crates/rt-api-native/src/router.rs
use std::convert::Infallible;

use axum::{
    body::{to_bytes, Body},
    extract::{DefaultBodyLimit, State},
    http::{header, HeaderMap, Request, StatusCode},
    middleware::{self, Next},
    response::{IntoResponse, Response},
    routing::{delete, get, post, put},
    Router,
};

use crate::{
    handlers::{
        add_torrent, add_torrent_peers, add_torrent_tags, apply_rss_rules, auth_login, auth_logout,
        auth_settings, bulk_action, cancel_job, categories, create_tag, cross_seed,
        delete_category, delete_saved_json, delete_tag, delete_torrent, diagnose_torrent,
        engine_commands, engine_diagnostics, get_torrent, get_user_agent, health, list_json_map,
        list_session_events, list_torrent_files, list_torrent_trackers, list_torrents,
        list_workflow_runs, live_torrent_stats, logs, metrics, patch_torrent_files,
        patch_torrent_trackers, pause_job, pause_torrent, reannounce_torrent, recheck_torrent,
        remove_torrent_tags, reset_auth_settings, restart_engine, resume_job, resume_torrent,
        rtorrent_settings, run_json_workflow, save_rtorrent_settings, session_features,
        session_settings, set_torrent_category, set_user_agent, sidebar_facets, storage,
        storage_execute_plan, storage_preview_plan, stream_events, tags, test_rss_rules,
        torrent_limits, tracker_health, transfer_info, transfer_limits, update_auth_settings,
        update_session_features, update_session_settings, update_torrent, update_torrent_limits,
        update_torrent_queue, update_transfer_limits, upsert_category, upsert_json_map,
    },
    state::AppState,
};
use rt_api_model::{
    api_token_allowed, bearer_token, csrf_request_allowed, has_browser_request_headers,
    has_session_cookie, request_fingerprint, valid_idempotency_key, CachedResponse,
    IdempotencyClaim, MAX_IDEMPOTENCY_BODY_BYTES,
};
use tower::limit::GlobalConcurrencyLimitLayer;

const MAX_NATIVE_LARGE_BODY_REQUESTS: usize = 4;
const MAX_NATIVE_DEFAULT_BODY_BYTES: usize = 8 * 1024 * 1024;
const MAX_NATIVE_TORRENT_BODY_BYTES: usize = 96 * 1024 * 1024;
const MAX_NATIVE_AUTH_BODY_BYTES: usize = 16 * 1024;

pub fn build_router(state: AppState) -> Router {
    let large_body_limit = GlobalConcurrencyLimitLayer::new(MAX_NATIVE_LARGE_BODY_REQUESTS);
    Router::new()
        .route("/health", get(health))
        .route("/metrics", get(metrics))
        .route(
            "/api/v1/auth/login",
            post(auth_login).layer(DefaultBodyLimit::max(MAX_NATIVE_AUTH_BODY_BYTES)),
        )
        .route("/api/v1/auth/logout", post(auth_logout))
        .route(
            "/api/v1/auth/settings",
            get(auth_settings)
                .put(update_auth_settings)
                .delete(reset_auth_settings)
                .layer(DefaultBodyLimit::max(MAX_NATIVE_AUTH_BODY_BYTES)),
        )
        .route(
            "/api/v1/torrents",
                get(list_torrents)
                .post(add_torrent)
                .layer::<_, Infallible>(large_body_limit)
                .layer(DefaultBodyLimit::max(MAX_NATIVE_TORRENT_BODY_BYTES)),
        )
        .route("/api/v1/torrents/live", get(live_torrent_stats))
        .route(
            "/api/v1/torrents/:hash",
            get(get_torrent).put(update_torrent).delete(delete_torrent),
        )
        .route(
            "/api/v1/torrents/:hash/stop",
            axum::routing::post(pause_torrent),
        )
        .route(
            "/api/v1/torrents/:hash/pause",
            axum::routing::post(pause_torrent),
        )
        .route(
            "/api/v1/torrents/:hash/start",
            axum::routing::post(resume_torrent),
        )
        .route(
            "/api/v1/torrents/:hash/resume",
            axum::routing::post(resume_torrent),
        )
        .route(
            "/api/v1/torrents/:hash/recheck",
            axum::routing::post(recheck_torrent),
        )
        .route(
            "/api/v1/torrents/:hash/reannounce",
            axum::routing::post(reannounce_torrent),
        )
        .route("/api/v1/torrents/:hash/category", put(set_torrent_category))
        .route(
            "/api/v1/torrents/:hash/limits",
            get(torrent_limits).put(update_torrent_limits),
        )
        .route("/api/v1/torrents/:hash/peers", post(add_torrent_peers))
        .route("/api/v1/torrents/queue", post(update_torrent_queue))
        .route(
            "/api/v1/transfer/limits",
            get(transfer_limits).put(update_transfer_limits),
        )
        .route("/api/v1/transfer/info", get(transfer_info))
        .route(
            "/api/v1/session/features",
            get(session_features).put(update_session_features),
        )
        .route(
            "/api/v1/session/settings",
            get(session_settings).patch(update_session_settings),
        )
        .route(
            "/api/v1/torrents/:hash/tags",
            post(add_torrent_tags)
                .patch(crate::handlers::patch_torrent_tags)
                .delete(remove_torrent_tags),
        )
        .route("/api/v1/categories", get(categories).post(upsert_category))
        .route("/api/v1/categories/:name", delete(delete_category))
        .route("/api/v1/tags", get(tags).post(create_tag))
        .route("/api/v1/tags/:name", delete(delete_tag))
        .route("/api/v1/bulk/:action", post(bulk_action))
        .route("/api/v1/cross-seed", post(cross_seed))
        .route("/api/v1/tracker-health", get(tracker_health))
        .route("/api/v1/sidebar-facets", get(sidebar_facets))
        .route("/api/v1/logs", get(logs))
        .route(
            "/api/v1/saved-views",
            get(list_json_map::<crate::handlers::SavedViewsStore>)
                .post(upsert_json_map::<crate::handlers::SavedViewsStore>),
        )
        .route(
            "/api/v1/saved-views/:id",
            delete(delete_saved_json::<crate::handlers::SavedViewsStore>),
        )
        .route(
            "/api/v1/ratio-groups",
            get(list_json_map::<crate::handlers::RatioGroupsStore>)
                .post(upsert_json_map::<crate::handlers::RatioGroupsStore>),
        )
        .route(
            "/api/v1/ratio-groups/:id",
            post(run_json_workflow::<crate::handlers::RatioGroupsStore>)
                .delete(delete_saved_json::<crate::handlers::RatioGroupsStore>),
        )
        .route(
            "/api/v1/workflows",
            get(list_json_map::<crate::handlers::WorkflowsStore>)
                .post(upsert_json_map::<crate::handlers::WorkflowsStore>),
        )
        .route(
            "/api/v1/workflows/:id",
            post(run_json_workflow::<crate::handlers::WorkflowsStore>)
                .delete(delete_saved_json::<crate::handlers::WorkflowsStore>),
        )
        .route("/api/v1/workflow-runs", get(list_workflow_runs))
        .route(
            "/api/v1/rss-rules",
            get(list_json_map::<crate::handlers::RssRulesStore>)
                .post(upsert_json_map::<crate::handlers::RssRulesStore>),
        )
        .route("/api/v1/rss-rules/test", post(test_rss_rules))
        .route("/api/v1/rss-rules/apply", post(apply_rss_rules))
        .route(
            "/api/v1/rss-rules/:id",
            delete(delete_saved_json::<crate::handlers::RssRulesStore>),
        )
        .route("/api/v1/engine", get(engine_diagnostics))
        .route("/api/v1/engine/commands", get(engine_commands))
        .route(
            "/api/v1/engine/rtorrent-settings",
            get(rtorrent_settings).put(save_rtorrent_settings),
        )
        .route("/api/v1/engine/restart", post(restart_engine))
        .route(
            "/api/v1/engine/user-agent",
            get(get_user_agent).put(set_user_agent),
        )
        .route(
            "/api/v1/settings/user-agent",
            get(get_user_agent).put(set_user_agent),
        )
        .route(
            "/api/v1/torrents/:hash/files",
            get(list_torrent_files).patch(patch_torrent_files),
        )
        .route(
            "/api/v1/torrents/:hash/trackers",
            get(list_torrent_trackers).patch(patch_torrent_trackers),
        )
        .route("/api/v1/torrents/:hash/diagnostics", get(diagnose_torrent))
        .route("/api/v1/events", get(stream_events))
        .route("/api/v1/session-events", get(list_session_events))
        .route("/api/v1/jobs", get(crate::handlers::list_jobs))
        .route("/api/v1/jobs/:id/pause", post(pause_job))
        .route("/api/v1/jobs/:id/resume", post(resume_job))
        .route("/api/v1/jobs/:id/cancel", post(cancel_job))
        .route("/api/v1/storage", get(storage))
        .route("/api/v1/storage/plan", post(storage_preview_plan))
        .route("/api/v1/storage/execute", post(storage_execute_plan))
        .route_layer(middleware::from_fn_with_state(
            state.clone(),
            torrentng_idempotency_guard,
        ))
        .route_layer(middleware::from_fn_with_state(state.clone(), torrentng_auth_guard))
        // Keep ordinary JSON/form routes well below the torrent upload limit.
        // The route-local limit above is the only path that needs the larger
        // TorrentNG JSON/base64 envelope.
        .layer(DefaultBodyLimit::max(MAX_NATIVE_DEFAULT_BODY_BYTES))
        .with_state(state)
}

/// Coalesce retries of successful TorrentNG HTTP mutations. Durable engine jobs
/// cover restart recovery; this bounded middleware covers the client timeout
/// window and rejects accidental reuse of a key for a different request.
async fn torrentng_idempotency_guard(
    State(state): State<AppState>,
    req: Request<Body>,
    next: Next,
) -> Response {
    if !matches!(
        req.method(),
        &axum::http::Method::POST
            | &axum::http::Method::PUT
            | &axum::http::Method::PATCH
            | &axum::http::Method::DELETE
    ) || torrentng_auth_path(req.uri().path())
    {
        return next.run(req).await;
    }

    let Some(key) = req
        .headers()
        .get("idempotency-key")
        .and_then(|value| value.to_str().ok())
        .map(str::to_owned)
    else {
        return next.run(req).await;
    };
    if !valid_idempotency_key(&key) {
        return (StatusCode::BAD_REQUEST, "invalid Idempotency-Key").into_response();
    }

    let (mut parts, body) = req.into_parts();
    let body = match to_bytes(body, MAX_IDEMPOTENCY_BODY_BYTES).await {
        Ok(body) => body,
        Err(_) => return StatusCode::PAYLOAD_TOO_LARGE.into_response(),
    };
    let fingerprint =
        request_fingerprint(parts.method.as_str(), &parts.uri.to_string(), body.as_ref());
    parts.headers.remove(header::CONTENT_LENGTH);
    let req = Request::from_parts(parts, Body::from(body.to_vec()));

    loop {
        match state.idempotency.claim(&key, fingerprint) {
            IdempotencyClaim::Execute => break,
            IdempotencyClaim::Wait(notify) => {
                notify.notified().await;
            }
            IdempotencyClaim::Replay(cached) => return replay_response(cached),
            IdempotencyClaim::Conflict => {
                return (
                    StatusCode::UNPROCESSABLE_ENTITY,
                    "Idempotency-Key was already used for a different request",
                )
                    .into_response();
            }
            IdempotencyClaim::Saturated => {
                return (
                    StatusCode::SERVICE_UNAVAILABLE,
                    "idempotency store is saturated; retry later",
                )
                    .into_response();
            }
        }
    }

    let mut execution = state.idempotency.execution_guard(&key, fingerprint);
    let response = next.run(req).await;
    let (parts, body) = response.into_parts();
    let body = match to_bytes(body, MAX_IDEMPOTENCY_BODY_BYTES).await {
        Ok(body) => body,
        Err(_) => {
            return (
                StatusCode::INTERNAL_SERVER_ERROR,
                "mutation response exceeded the idempotency response limit",
            )
                .into_response();
        }
    };
    let response = Response::from_parts(parts.clone(), Body::from(body.clone()));
    if parts.status.is_success() {
        let headers = parts
            .headers
            .iter()
            .map(|(name, value)| (name.to_string(), value.as_bytes().to_vec()))
            .collect();
        execution.complete(CachedResponse {
            status: parts.status.as_u16(),
            headers,
            body: body.to_vec(),
        });
    } else {
        execution.abandon();
    }
    response
}

fn replay_response(cached: CachedResponse) -> Response {
    // Keep the bound explicit at the final user-controlled allocation site as
    // well as at request capture. This protects against corrupted/injected
    // in-memory cache entries and makes the idempotency replay contract
    // auditable by static analysis.
    if cached.body.len() > MAX_IDEMPOTENCY_BODY_BYTES {
        return (
            StatusCode::INTERNAL_SERVER_ERROR,
            "cached idempotency response exceeded the replay limit",
        )
            .into_response();
    }
    let mut response = Response::new(Body::from(cached.body));
    *response.status_mut() = StatusCode::from_u16(cached.status).unwrap_or(StatusCode::OK);
    for (name, value) in cached.headers {
        let Ok(name) = axum::http::HeaderName::from_bytes(name.as_bytes()) else {
            continue;
        };
        let Ok(value) = axum::http::HeaderValue::from_bytes(&value) else {
            continue;
        };
        response.headers_mut().append(name, value);
    }
    response.headers_mut().insert(
        axum::http::HeaderName::from_static("idempotency-replayed"),
        axum::http::HeaderValue::from_static("true"),
    );
    response
}

async fn torrentng_auth_guard(
    State(state): State<AppState>,
    req: Request<Body>,
    next: Next,
) -> Response {
    let path = req.uri().path();
    if bearer_token(req.headers()).is_some_and(|token| api_token_allowed(&state.api_tokens, &token))
    {
        return next.run(req).await;
    }

    if torrentng_public_path(path) {
        if torrentng_auth_path(path)
            && is_mutating_request(&req)
            && has_browser_request_headers(req.headers())
            && !csrf_request_allowed(req.headers())
        {
            return (
                StatusCode::FORBIDDEN,
                "cross-origin browser authentication request rejected",
            )
                .into_response();
        }
        return next.run(req).await;
    }

    if state.api_tokens.is_empty() {
        if let Some(session_token) = state.local_webui_session_token.as_ref() {
            let session_valid = torrentng_presented_token(req.headers()).is_some_and(|token| {
                api_token_allowed(std::slice::from_ref(session_token), &token)
            });
            if path == "/api/v1/auth/settings" && !session_valid {
                return (
                    StatusCode::UNAUTHORIZED,
                    [(header::CONTENT_TYPE, "application/json")],
                    r#"{"code":"UNAUTHORIZED","message":"login required"}"#,
                )
                    .into_response();
            }
            if has_browser_request_headers(req.headers()) {
                if !session_valid {
                    return (
                        StatusCode::UNAUTHORIZED,
                        [(header::CONTENT_TYPE, "application/json")],
                        r#"{"code":"UNAUTHORIZED","message":"login required"}"#,
                    )
                        .into_response();
                }
                if has_session_cookie(req.headers(), &["tng_session"])
                    && is_mutating_request(&req)
                    && !csrf_request_allowed(req.headers())
                {
                    return (StatusCode::FORBIDDEN, "cross-site cookie mutation rejected")
                        .into_response();
                }
            }
            return next.run(req).await;
        }
        if is_mutating_request(&req)
            && has_browser_request_headers(req.headers())
            && !csrf_request_allowed(req.headers())
        {
            return (
                StatusCode::FORBIDDEN,
                "cross-origin browser request rejected",
            )
                .into_response();
        }
        return next.run(req).await;
    }

    if torrentng_presented_token(req.headers())
        .is_some_and(|token| api_token_allowed(&state.api_tokens, &token))
    {
        if has_session_cookie(req.headers(), &["tng_session"])
            && is_mutating_request(&req)
            && !csrf_request_allowed(req.headers())
        {
            return (StatusCode::FORBIDDEN, "cross-site cookie mutation rejected").into_response();
        }
        return next.run(req).await;
    }

    (
        StatusCode::UNAUTHORIZED,
        [(header::CONTENT_TYPE, "application/json")],
        r#"{"code":"UNAUTHORIZED","message":"missing or invalid API token"}"#,
    )
        .into_response()
}

fn is_mutating_request(req: &Request<Body>) -> bool {
    matches!(
        *req.method(),
        axum::http::Method::POST
            | axum::http::Method::PUT
            | axum::http::Method::PATCH
            | axum::http::Method::DELETE
    )
}

fn torrentng_public_path(path: &str) -> bool {
    matches!(
        path,
        "/health" | "/api/v1/auth/login" | "/api/v1/auth/logout"
    )
}

fn torrentng_auth_path(path: &str) -> bool {
    matches!(path, "/api/v1/auth/login" | "/api/v1/auth/logout")
}

fn torrentng_presented_token(headers: &HeaderMap) -> Option<String> {
    bearer_token(headers).or_else(|| {
        headers
            .get(header::COOKIE)
            .and_then(|value| value.to_str().ok())
            .and_then(torrentng_session_cookie)
    })
}

fn torrentng_session_cookie(cookie: &str) -> Option<String> {
    cookie.split(';').find_map(|part| {
        let part = part.trim();
        part.strip_prefix("tng_session=")
            .and_then(cookie_component_decode)
    })
}

fn cookie_component_decode(input: &str) -> Option<String> {
    let bytes = input.as_bytes();
    let mut output = Vec::with_capacity(bytes.len());
    let mut index = 0;
    while index < bytes.len() {
        if bytes[index] != b'%' {
            output.push(bytes[index]);
            index += 1;
            continue;
        }
        if index + 2 >= bytes.len() {
            return None;
        }
        let high = hex_value(bytes[index + 1])?;
        let low = hex_value(bytes[index + 2])?;
        output.push((high << 4) | low);
        index += 3;
    }
    String::from_utf8(output).ok()
}

fn hex_value(byte: u8) -> Option<u8> {
    match byte {
        b'0'..=b'9' => Some(byte - b'0'),
        b'a'..=b'f' => Some(byte - b'a' + 10),
        b'A'..=b'F' => Some(byte - b'A' + 10),
        _ => None,
    }
}

#[cfg(test)]
mod tests {
    use super::{build_router, torrentng_auth_path, torrentng_public_path};
    use crate::state::AppState;
    use axum::{body::Body, http::Request};
    use std::sync::Arc;
    use tower::ServiceExt;

    #[test]
    fn native_auth_paths_are_exact() {
        assert!(torrentng_auth_path("/api/v1/auth/login"));
        assert!(torrentng_auth_path("/api/v1/auth/logout"));
        assert!(!torrentng_auth_path("/api/v1/nested/auth/login"));
        assert!(!torrentng_auth_path("/api/v1/auth/logout/extra"));
    }

    #[test]
    fn native_public_paths_do_not_include_future_auth_routes() {
        assert!(torrentng_public_path("/health"));
        assert!(torrentng_public_path("/api/v1/auth/login"));
        assert!(torrentng_public_path("/api/v1/auth/logout"));
        assert!(!torrentng_public_path("/api/v1/auth/refresh"));
        assert!(!torrentng_public_path("/api/v1/torrents"));
    }

    #[tokio::test]
    async fn native_browser_login_rejects_cross_origin_cookie_seeding() {
        let mut state = AppState::new();
        state.api_tokens = Arc::new(vec!["secret".to_owned()]);
        let app = build_router(state);
        let response = app
            .oneshot(
                Request::builder()
                    .method("POST")
                    .uri("/api/v1/auth/login")
                    .header("host", "torrentng.example")
                    .header("origin", "https://attacker.example")
                    .header("content-type", "application/x-www-form-urlencoded")
                    .body(Body::from("token=secret"))
                    .unwrap(),
            )
            .await
            .unwrap();
        assert_eq!(response.status(), axum::http::StatusCode::FORBIDDEN);
    }
}