# docs/API.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TorrentNG API",
    "version": "0.1.0",
    "description": "Machine-readable contract for the implemented TorrentNG /api/v1 JSON REST surface. Compatibility facades are documented by their client-specific matrices. Mutating requests accept an optional Idempotency-Key; successful responses may be replayed with Idempotency-Replayed: true. API tokens are bearer tokens or tng_session cookies."
  },
  "servers": [{"url": "/"}],
  "security": [{"bearerAuth": []}, {"sessionCookie": []}],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {"operationId": "authLogin", "security": [], "requestBody": {"$ref": "#/components/requestBodies/AuthLoginForm"}, "responses": {"200": {"description": "Login accepted; sets the tng_session cookie", "content": {"text/plain": {"schema": {"type": "string", "enum": ["Ok."]}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "429": {"description": "Login rate limit reached; response includes Retry-After"}}}
    },
...
    "/api/v1/auth/settings": {
      "get": {"operationId": "getAuthSettings", "responses": {"200": {"description": "Current safe authentication settings; does not return the password", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthSettings"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}}},
      "put": {"operationId": "updateAuthSettings", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/UpdateAuthSettings"}, "responses": {"200": {"description": "Updated safe authentication settings", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthSettings"}}}}, "400": {"$ref": "#/components/responses/JsonError"}, "401": {"$ref": "#/components/responses/Unauthorized"}}},
      "delete": {"operationId": "resetAuthSettings", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"200": {"description": "Config-file authentication settings restored", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthSettings"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}}}
    },
    "/api/v1/torrents": {
      "get": {"operationId": "listTorrents", "parameters": [{"$ref": "#/components/parameters/Limit"}, {"$ref": "#/components/parameters/Offset"}, {"$ref": "#/components/parameters/Snapshot"}, {"$ref": "#/components/parameters/Filter"}], "responses": {"200": {"$ref": "#/components/responses/JsonOk"}, "401": {"$ref": "#/components/responses/Unauthorized"}}},
      "post": {"operationId": "addTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/AddTorrent"}, "responses": {"201": {"$ref": "#/components/responses/JsonOk"}, "400": {"$ref": "#/components/responses/JsonError"}, "413": {"$ref": "#/components/responses/PayloadTooLarge"}, "422": {"$ref": "#/components/responses/InvalidRequest"}}}
    },
...
    "/api/v1/torrents/{hash}": {
      "parameters": [{"$ref": "#/components/parameters/Hash"}],
      "get": {"operationId": "getTorrent", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}, "404": {"$ref": "#/components/responses/NotFound"}}},
      "put": {"operationId": "updateTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "202": {"$ref": "#/components/responses/JsonOk"}, "400": {"$ref": "#/components/responses/JsonError"}, "404": {"$ref": "#/components/responses/NotFound"}}},
      "delete": {"operationId": "deleteTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"$ref": "#/components/responses/Empty"}, "202": {"$ref": "#/components/responses/JsonOk"}, "404": {"$ref": "#/components/responses/NotFound"}}}
    },
    "/api/v1/torrents/{hash}/start": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "post": {"operationId": "startTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"$ref": "#/components/responses/Empty"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/torrents/{hash}/resume": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "post": {"operationId": "resumeTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"$ref": "#/components/responses/Empty"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/torrents/{hash}/stop": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "post": {"operationId": "stopTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"$ref": "#/components/responses/Empty"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/torrents/{hash}/pause": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "post": {"operationId": "pauseTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"$ref": "#/components/responses/Empty"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/torrents/{hash}/recheck": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "post": {"operationId": "recheckTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"202": {"$ref": "#/components/responses/JsonOk"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/torrents/{hash}/reannounce": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "post": {"operationId": "reannounceTorrent", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"$ref": "#/components/responses/Empty"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/torrents/{hash}/category": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "put": {"operationId": "setTorrentCategory", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/SetTorrentCategory"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/v1/torrents/{hash}/limits": {
      "parameters": [{"$ref": "#/components/parameters/Hash"}],
      "get": {"operationId": "getTorrentLimits", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}, "404": {"$ref": "#/components/responses/NotFound"}}},
      "put": {"operationId": "updateTorrentLimits", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "501": {"$ref": "#/components/responses/JsonError"}}}
    },
    "/api/v1/torrents/{hash}/peers": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "post": {"operationId": "addTorrentPeers", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}}}},
    "/api/v1/torrents/queue": {"post": {"operationId": "updateTorrentQueue", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}}}},
    "/api/v1/torrents/{hash}/tags": {
      "parameters": [{"$ref": "#/components/parameters/Hash"}],
      "post": {"operationId": "addTorrentTags", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/TorrentTags"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/InvalidRequest"}}},
      "patch": {"operationId": "patchTorrentTags", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/PatchTorrentTags"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/InvalidRequest"}}},
      "delete": {"operationId": "removeTorrentTags", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/TorrentTags"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/InvalidRequest"}}}
    },
    "/api/v1/torrents/{hash}/files": {
      "parameters": [{"$ref": "#/components/parameters/Hash"}],
      "get": {"operationId": "listTorrentFiles", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}}},
      "patch": {"operationId": "patchTorrentFiles", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}}}
    },
    "/api/v1/torrents/{hash}/trackers": {
      "parameters": [{"$ref": "#/components/parameters/Hash"}],
      "get": {"operationId": "listTorrentTrackers", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}}},
      "patch": {"operationId": "patchTorrentTrackers", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}}}
    },
    "/api/v1/torrents/{hash}/diagnostics": {"parameters": [{"$ref": "#/components/parameters/Hash"}], "get": {"operationId": "diagnoseTorrent", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}}}},
    "/api/v1/transfer/info": {"get": {"operationId": "transferInfo", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}, "503": {"$ref": "#/components/responses/JsonError"}}}},
    "/api/v1/transfer/limits": {
      "get": {"operationId": "transferLimits", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}}},
      "put": {"operationId": "updateTransferLimits", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}}}
    },
    "/api/v1/session/features": {
      "get": {"operationId": "sessionFeatures", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}}},
      "put": {"operationId": "updateSessionFeatures", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}}}
    },
    "/api/v1/session/settings": {
      "get": {"operationId": "sessionSettings", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}, "503": {"$ref": "#/components/responses/JsonError"}}},
      "patch": {"operationId": "updateSessionSettings", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}, "503": {"$ref": "#/components/responses/JsonError"}}}
    },
    "/api/v1/categories": {
      "get": {"operationId": "listCategories", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}}},
      "post": {"operationId": "upsertCategory", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/CategoryUpsert"}, "responses": {"200": {"$ref": "#/components/responses/JsonOk"}, "400": {"$ref": "#/components/responses/JsonError"}, "422": {"$ref": "#/components/responses/InvalidRequest"}}}
    },
    "/api/v1/categories/{name}": {"parameters": [{"$ref": "#/components/parameters/Name"}], "delete": {"operationId": "deleteCategory", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}}}},
    "/api/v1/tags": {
      "get": {"operationId": "listTags", "responses": {"200": {"$ref": "#/components/responses/JsonOk"}}},
      "post": {"operationId": "createTag", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/TagCreate"}, "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}, "422": {"$ref": "#/components/responses/InvalidRequest"}}}
    },
    "/api/v1/tags/{name}": {"parameters": [{"$ref": "#/components/parameters/Name"}], "delete": {"operationId": "deleteTag", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"$ref": "#/components/responses/Empty"}, "400": {"$ref": "#/components/responses/JsonError"}}}},
    "/api/v1/bulk/{action}": {"parameters": [{"$ref": "#/components/parameters/Action"}], "post": {"operationId": "bulkAction", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/BulkAction"}, "responses": {"200": {"$ref": "#/components/responses/JsonOk"}, "400": {"$ref": "#/components/responses/JsonError"}, "413": {"$ref": "#/components/responses/PayloadTooLarge"}, "422": {"$ref": "#/components/responses/InvalidRequest"}}}},
    "/api/v1/cross-seed": {"post": {"operationId": "crossSeed", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"$ref": "#/components/requestBodies/JsonObject"}, "responses": {"200": {"$ref": "#/components/responses/JsonOk"}}}},
...
    "/api/v1/engine/restart": {"post": {"operationId": "restartEngine", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"501": {"$ref": "#/components/responses/JsonError"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer"},
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "tng_session"}
    },
...
    "requestBodies": {
      "AuthLoginForm": {"required": true, "description": "Supply configured username/password or an API token in either field.", "content": {"application/x-www-form-urlencoded": {"schema": {"type": "object", "properties": {"username": {"type": "string"}, "password": {"type": "string", "format": "password"}}}}}},
...
    },
    "responses": {
...
      "Unauthorized": {"description": "Missing or invalid API token"},
...
      "Empty": {"description": "Mutation accepted with no response body"}
    },
    "schemas": {
...
      "AuthSettings": {"type": "object", "required": ["username", "password_is_default", "default_credentials_allowed", "api_token_login_enabled"], "properties": {"username": {"type": "string"}, "password_is_default": {"type": "boolean"}, "default_credentials_allowed": {"type": "boolean"}, "api_token_login_enabled": {"type": "boolean"}}, "additionalProperties": false},
...
    }
  }
}